Reorganize repo to reflect changed to libari
I just reorganized this weird repository to somthing more concrete.
Before, this repository was just a thrown together project with very
little organization, I am renaming this repository from DataStructures
to libari because most of these classes are in the namespace ari.

Please see the README.md for more info